+++ qualtran/bloqs/factoring/ecc/ec_phase_estimate_r.py
@@ -33,9 +33,9 @@ from qualtran import (
 from qualtran.bloqs.basic_gates import PlusState
 from qualtran.resource_counting import BloqCountT, SympySymbolAllocator
 
-from ._ec_add_r import ECAddR
-from ._ec_point import ECPoint
 from ._ecc_shims import MeasureQFT
+from .ec_add_r import ECAddR
+from .ec_point import ECPoint
 
 
 @frozen
